Fix bayes search and split up sweep algorithms
